Add gutter, moved view resources, settings and gutter to sub modules
- F1, F2 changes dark / light theme (albeit crappily)
- Adds line numbers (albeit crappy)
- Allowd text groups to have multiline handled layouts
- Allows primitives to change colour during lifetime

+++ eddy/src/actions.rs
@@ -42,7 +42,7 @@ pub enum Action {
         col: u64,
         ty: GestureType,
     },
-    Scroll(Range<i64>),
+    Scroll(Range<usize>),
     Repeat(Box<Action>, Quantity),
 }
